requirements.txt
asgiref==3.6.0
attrs==23.1.0
colorama==0.4.6
Django==4.2.1
django-admin-autocomplete-filter==0.7.1
django-cors-headers==4.0.0
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
execnet==2.0.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
model-bakery==1.11.0
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
pydot==2.0.0
pyparsing==3.1.2
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytest-xdist==3.5.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
